fix: remove duplicate next-shims.d.ts from fixtures
